+++ examples/keyboard-navigation/webpack.config.babel.js
@@ -5,31 +5,36 @@ const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 const sourcePath = path.join(__dirname, 'src');
 
 module.exports = {
-  entry: path.resolve(sourcePath, 'App.js'),
+  entry: path.resolve(sourcePath, 'Keyboard.js'),
   output: {
-    path: __dirname,
-    filename: 'bundle.js',
+     path: __dirname,
+     filename: 'bundle.js'
   },
   resolve: {
     extensions: ['.js', '.jsx'],
-    modules: [sourcePath, path.resolve(__dirname, 'node_modules')],
+    modules: [
+      sourcePath,
+      path.resolve(__dirname, 'node_modules')
+    ]
   },
   module: {
     rules: [
       {
         test: /\.(js|jsx)$/,
         exclude: /node_modules/,
-        use: ['babel-loader'],
-        include: sourcePath,
-      },
-    ],
+        use: [
+          'babel-loader'
+        ],
+        include: sourcePath
+      }
+    ]
   },
   plugins: [
     new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify('production'),
+      'process.env.NODE_ENV': JSON.stringify('production')
     }),
     new webpack.optimize.ModuleConcatenationPlugin(),
     new UglifyJsPlugin(),
     new webpack.HashedModuleIdsPlugin(),
-  ],
-};
+  ]
+}